requirements.txt
numpy>=1.21.0,<2.0.0
scikit-learn>=1.2.2
matplotlib>=3.5.2
seaborn>=0.11.2
jsonschema>=4.17.3
tqdm>=4.64.0
joblib>=1.2.0
polars>=0.17.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
hydra-core==1.3.2
omegaconf==2.3.0
mlflow>=1.28.0,<3.0.0
captum>=0.5.0
optuna>=3.0.0
pytorch-lightning>=2.0.0
pytest>=7.0.0
tensorboard
